+++ src/Core/Validation/Validators/JsonValidator.php
@@ -50,6 +50,7 @@ class JsonValidator extends BaseValidator
             Version::V_1_1 => null, // unsupported version
             Version::V_1_2 => Resources::FILE_CDX_JSON_SCHEMA_1_2,
             Version::V_1_3 => Resources::FILE_CDX_JSON_SCHEMA_1_3,
+            Version::V_1_4 => Resources::FILE_CDX_JSON_SCHEMA_1_4,
         ];
     }