+++ tests/Core/SerializeToXmlTest.php
@@ -28,6 +28,7 @@ use CycloneDX\Core\Serialize\XmlSerializer;
 use CycloneDX\Core\Spec\Spec11;
 use CycloneDX\Core\Spec\Spec12;
 use CycloneDX\Core\Spec\Spec13;
+use CycloneDX\Core\Spec\Spec14;
 use CycloneDX\Core\Validation\Validators\XmlValidator;
 use PHPUnit\Framework\TestCase;
 
@@ -111,4 +112,28 @@ class SerializeToXmlTest extends TestCase
     }
 
     // endregion Spec 1.3
+
+    // region Spec 1.4
+
+    /**
+     * This test might be slow.
+     * This test might require online-connectivity.
+     *
+     * @dataProvider \CycloneDX\Tests\_data\BomModelProvider::allBomTestData
+     *
+     * @throws \Exception on validation failure
+     */
+    public function testSchema14(Bom $bom): void
+    {
+        $spec = new Spec14();
+        $serializer = new XmlSerializer($spec);
+        $validator = new XmlValidator($spec);
+
+        $xml = $serializer->serialize($bom);
+        $validationErrors = $validator->validateString($xml);
+
+        self::assertNull($validationErrors);
+    }
+
+    // endregion Spec 1.4
 }